Moved SimpleInjector IPackage extensions methods to Api project
