+++ src/RepoM.Api/Plugins/SimpleInjector/README.txt
@@ -1,3 +1,2 @@
 Copied code from SimpleInjector repository in order to fix this issue: https://github.com/simpleinjector/SimpleInjector/issues/964
-
-When new package arrives, this code will be removed and the package reference will be used again.
+Also changed IPackage interface for RepoM needs. So it will not be possible to use original package.
