+++ src/RepoM.App/Bootstrapper.cs
@@ -46,6 +46,7 @@ using SimpleInjector;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
 using RepoM.Api.Plugins;
+using RepoM.Api.Plugins.SimpleInjector;
 using RepoM.App.Plugins;
 using RepoM.App.Services.HotKey;
