+++ tests/RepoM.Plugin.LuceneQueryParser.Tests/LuceneQueryParserPackageTest.cs
@@ -68,10 +68,10 @@ public class LuceneQueryParserPackageTest
     }
 }
 
-file static class ExtensionIPackageWithConfiguration
+file static class PackageExtensions
 {
     // tmp for fixing tests.
-    public static void RegisterServices(this IPackageWithConfiguration self, Container container)
+    public static void RegisterServices(this IPackage self, Container container)
     {
         self.RegisterServicesAsync(container, null!).GetAwaiter().GetResult();
     }
