+++ tests/RepoM.Plugin.WindowsExplorerGitInfo.Tests/WindowsExplorerGitInfoPackageTests.cs
@@ -45,10 +45,10 @@ public class WindowsExplorerGitInfoPackageTests
     }
 }
 
-file static class ExtensionIPackageWithConfiguration
+file static class PackageExtensions
 {
     // tmp for fixing tests.
-    public static void RegisterServices(this IPackageWithConfiguration self, Container container)
+    public static void RegisterServices(this IPackage self, Container container)
     {
         self.RegisterServicesAsync(container, null!).GetAwaiter().GetResult();
     }
